move run_program_and_adapter to dev utils
