+++ stwo_cairo_prover/crates/dev_utils/src/bin/prove_from_compiled_program.rs
@@ -2,8 +2,8 @@ use std::path::PathBuf;
 
 use cairo_air::utils::ProofFormat;
 use clap::Parser;
-use dev_utils::utils::{create_and_serialize_proof, Error};
-use stwo_cairo_adapter::test_utils::{read_compiled_cairo_program, run_program_and_adapter};
+use dev_utils::utils::{create_and_serialize_proof, run_program_and_adapter, Error};
+use stwo_cairo_adapter::test_utils::read_compiled_cairo_program;
 use tracing::{span, Level};
 use tracing_subscriber::fmt::format::FmtSpan;
