+++ stwo_cairo_prover/crates/prover/src/prover.rs
@@ -179,7 +179,8 @@ pub fn default_prod_prover_parameters() -> ProverParameters {
 #[cfg(test)]
 pub mod tests {
     use cairo_air::preprocessed::testing_preprocessed_tree;
-    use stwo_cairo_adapter::test_utils::{get_test_program, run_program_and_adapter};
+    use dev_utils::utils::run_program_and_adapter;
+    use stwo_cairo_adapter::test_utils::get_test_program;
 
     use crate::debug_tools::assert_constraints::assert_cairo_constraints;
     #[test]
@@ -439,7 +440,7 @@ pub mod tests {
 
         /// These tests' inputs were generated using cairo-vm with 50 instances of each builtin.
         pub mod builtin_tests {
-            use stwo_cairo_adapter::test_utils::run_program_and_adapter;
+            use dev_utils::utils::run_program_and_adapter;
             use test_log::test;
 
             use super::*;
